+++ core/src/main/java/com/github/shadowsocks/net/DnsResolverCompat.kt
@@ -20,18 +20,14 @@
 
 package com.github.shadowsocks.net
 
-import android.annotation.SuppressLint
 import android.annotation.TargetApi
 import android.net.DnsResolver
 import android.net.Network
 import android.os.Build
 import android.os.CancellationSignal
-import android.system.ErrnoException
 import com.github.shadowsocks.Core
-import com.github.shadowsocks.utils.int
 import kotlinx.coroutines.*
 import org.xbill.DNS.*
-import java.io.FileDescriptor
 import java.io.IOException
 import java.net.Inet4Address
 import java.net.Inet6Address
@@ -47,12 +43,10 @@ sealed class DnsResolverCompat {
             when (Build.VERSION.SDK_INT) {
                 in 29..Int.MAX_VALUE -> DnsResolverCompat29
                 in 23 until 29 -> DnsResolverCompat23
-                in 21 until 23 -> DnsResolverCompat21()
                 else -> error("Unsupported API level")
             }
         }
 
-        override fun bindSocket(network: Network, socket: FileDescriptor) = instance.bindSocket(network, socket)
         override suspend fun resolve(network: Network, host: String) = instance.resolve(network, host)
         override suspend fun resolveOnActiveNetwork(host: String) = instance.resolveOnActiveNetwork(host)
         override suspend fun resolveRaw(network: Network, query: ByteArray) = instance.resolveRaw(network, query)
@@ -74,29 +68,12 @@ sealed class DnsResolverCompat {
         }
     }
 
-    @Throws(IOException::class)
-    abstract fun bindSocket(network: Network, socket: FileDescriptor)
     abstract suspend fun resolve(network: Network, host: String): Array<InetAddress>
     abstract suspend fun resolveOnActiveNetwork(host: String): Array<InetAddress>
     abstract suspend fun resolveRaw(network: Network, query: ByteArray): ByteArray
     abstract suspend fun resolveRawOnActiveNetwork(query: ByteArray): ByteArray
 
-    @SuppressLint("PrivateApi")
-    private open class DnsResolverCompat21 : DnsResolverCompat() {
-        private val bindSocketToNetwork by lazy {
-            Class.forName("android.net.NetworkUtils").getDeclaredMethod(
-                    "bindSocketToNetwork", Int::class.java, Int::class.java)
-        }
-        private val netId by lazy { Network::class.java.getDeclaredField("netId") }
-        @SuppressLint("NewApi")
-        override fun bindSocket(network: Network, socket: FileDescriptor) {
-            val netId = netId.get(network)!!
-            val err = bindSocketToNetwork.invoke(null, socket.int, netId) as Int
-            if (err == 0) return
-            val message = "Binding socket to network $netId"
-            throw ErrnoException(message, -err).rethrowAsSocketException()
-        }
-
+    private object DnsResolverCompat23 : DnsResolverCompat() {
         /**
          * This dispatcher is used for noncancellable possibly-forever-blocking operations in network IO.
          *
@@ -108,9 +85,9 @@ sealed class DnsResolverCompat {
         }
 
         override suspend fun resolve(network: Network, host: String) =
-                GlobalScope.async(unboundedIO) { network.getAllByName(host) }.await()
+                withContext(unboundedIO) { network.getAllByName(host) }
         override suspend fun resolveOnActiveNetwork(host: String) =
-                GlobalScope.async(unboundedIO) { InetAddress.getAllByName(host) }.await()
+                withContext(unboundedIO) { InetAddress.getAllByName(host) }
 
         private suspend fun resolveRaw(query: ByteArray, networkSpecified: Boolean = true,
                                        hostResolver: suspend (String) -> Array<InetAddress>): ByteArray {
@@ -128,16 +105,18 @@ sealed class DnsResolverCompat {
                 Type.A -> false
                 Type.AAAA -> true
                 Type.PTR -> {
-                    // Android does not provide a PTR lookup API for Network prior to Android 10
+                    /* Android does not provide a PTR lookup API for Network prior to Android 10 */
                     if (networkSpecified) throw IOException(UnsupportedOperationException("Network unspecified"))
                     val ip = try {
                         ReverseMap.fromName(question.name)
                     } catch (e: IOException) {
                         throw UnsupportedOperationException(e)  // unrecognized PTR name
                     }
-                    val hostname = Name.fromString(GlobalScope.async(unboundedIO) { ip.hostName }.await())
+                    val hostname = withContext(unboundedIO) { ip.hostName }.let { hostname ->
+                        if (hostname == ip.hostAddress) null else Name.fromString("$hostname.")
+                    }
                     return prepareDnsResponse(request).apply {
-                        addRecord(PTRRecord(question.name, DClass.IN, TTL, hostname), Section.ANSWER)
+                        hostname?.let { addRecord(PTRRecord(question.name, DClass.IN, TTL, it), Section.ANSWER) }
                     }.toWire()
                 }
                 else -> throw UnsupportedOperationException("Unsupported query type $type")
@@ -159,19 +138,12 @@ sealed class DnsResolverCompat {
                 resolveRaw(query, false, this::resolveOnActiveNetwork)
     }
 
-    @TargetApi(23)
-    private object DnsResolverCompat23 : DnsResolverCompat21() {
-        override fun bindSocket(network: Network, socket: FileDescriptor) = network.bindSocket(socket)
-    }
-
     @TargetApi(29)
     private object DnsResolverCompat29 : DnsResolverCompat(), Executor {
         /**
          * This executor will run on its caller directly. On Q beta 3 thru 4, this results in calling in main thread.
          */
         override fun execute(command: Runnable) = command.run()
-
-        override fun bindSocket(network: Network, socket: FileDescriptor) = network.bindSocket(socket)
 
         private val activeNetwork get() = Core.connectivity.activeNetwork ?: throw IOException("no network")